+++ contracts/cw-native-staked-balance-voting/src/tests.rs
@@ -1,7 +1,10 @@
+use crate::contract::{migrate, CONTRACT_NAME, CONTRACT_VERSION};
 use crate::msg::{
-    ExecuteMsg, InstantiateMsg, ListStakersResponse, Owner, QueryMsg, StakerBalanceResponse,
+    ExecuteMsg, InstantiateMsg, ListStakersResponse, MigrateMsg, Owner, QueryMsg,
+    StakerBalanceResponse,
 };
 use crate::state::Config;
+use cosmwasm_std::testing::{mock_dependencies, mock_env};
 use cosmwasm_std::{coins, Addr, Coin, Empty, Uint128};
 use cw_controllers::ClaimsResponse;
 use cw_core_interface::voting::{
@@ -969,4 +972,14 @@ fn test_query_list_stakers() {
         .unwrap();
 
     assert_eq!(stakers, ListStakersResponse { stakers: vec![] });
+}
+
+#[test]
+pub fn test_migrate_update_version() {
+    let mut deps = mock_dependencies();
+    cw2::set_contract_version(&mut deps.storage, "my-contract", "old-version").unwrap();
+    migrate(deps.as_mut(), mock_env(), MigrateMsg {}).unwrap();
+    let version = cw2::get_contract_version(&deps.storage).unwrap();
+    assert_eq!(version.version, CONTRACT_VERSION);
+    assert_eq!(version.contract, CONTRACT_NAME);
 }
