+++ contracts/cw20-stake-external-rewards/src/contract.rs
@@ -13,7 +13,6 @@ use crate::ContractError::{
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
 
-use cosmwasm_std::testing::{mock_env, mock_dependencies};
 use cosmwasm_std::{
     from_binary, to_binary, Addr, BankMsg, Binary, Coin, CosmosMsg, Deps, DepsMut, Empty, Env,
     MessageInfo, Response, StdError, StdResult, Uint128, Uint256, WasmMsg,
@@ -515,9 +514,17 @@ pub fn query_pending_rewards(
 mod tests {
     use std::borrow::BorrowMut;
 
-    use crate::ContractError;
+    use crate::{
+        contract::{migrate, CONTRACT_NAME, CONTRACT_VERSION},
+        msg::MigrateMsg,
+        ContractError,
+    };
 
-    use cosmwasm_std::{coin, to_binary, Addr, Empty, Uint128};
+    use cosmwasm_std::{
+        coin,
+        testing::{mock_dependencies, mock_env},
+        to_binary, Addr, Empty, Uint128,
+    };
     use cw20::{Cw20Coin, Cw20ExecuteMsg, Denom};
     use cw_utils::Duration;
 
@@ -2195,14 +2202,14 @@ mod tests {
             .execute_contract(admin, reward_addr, &fund_msg, &reward_funding)
             .unwrap_err();
     }
-}
 
-#[test]
-pub fn test_migrate_update_version() {
-    let mut deps = mock_dependencies();
-    cw2::set_contract_version(&mut deps.storage, "my-contract", "old-version").unwrap();
-    migrate(deps.as_mut(), mock_env(), MigrateMsg {}).unwrap();
-    let version = cw2::get_contract_version(&deps.storage).unwrap();
-    assert_eq!(version.version, CONTRACT_VERSION);
-    assert_eq!(version.contract, CONTRACT_NAME);
+    #[test]
+    pub fn test_migrate_update_version() {
+        let mut deps = mock_dependencies();
+        cw2::set_contract_version(&mut deps.storage, "my-contract", "old-version").unwrap();
+        migrate(deps.as_mut(), mock_env(), MigrateMsg {}).unwrap();
+        let version = cw2::get_contract_version(&deps.storage).unwrap();
+        assert_eq!(version.version, CONTRACT_VERSION);
+        assert_eq!(version.contract, CONTRACT_NAME);
+    }
 }
